Move pytest configuration into pytest.ini.
We had placed our pytest configuration in setup.cfg, which is pretty
common, but deprecated. This isn't really a major problem in practice
for pytest itself (it's been deprecated for over 5 years yet still
supported), but the recent addition to the log format configuration
actually broke building packages on Python 2.7, as the format strings
could not be expanded to anything.

There are two places this could conceivably live: `pytest.ini` or
`pyproject.toml`. `pyproject.toml` is increasingly the future of
Python package configuration, but unsupported by pytest prior to 6.0,
which means we can't use it for Python 2.7.

So this change moves the configuration into `pytest.ini` instead. This
is universally supported and has highest precedence.

Testing Done:
Successfully ran pytest on all supported versions of Python.

Successfully built packages.

Reviewed at https://reviews.reviewboard.org/r/12264/