Set an explicit title for the RBTools documentation.
We were using the default Sphinx title of "RBTools <version>
documentation", which tended to have alpha version information in it.
This is really messy. We now use an explicit title of the form:
"RBTools X.Y Documentation".

Testing Done:
Built the docs and checked the resulting title on the pages.

Reviewed at https://reviews.reviewboard.org/r/12266/