Fix repository type detection when subversion isn't installed.
When `svn` is unavailable, the subversion client was erroring out during
repository type detection. This could be worked around by manually
specifying the repository type in the config, but that's obviously not
optimal. This change adds checks to ensure we don't try to do anything
if we can't run `svn`.

Testing Done:
- Tested rbtools when `svn` was not installed.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12274/