Fix ClearCase file:vob checking on Windows.
In the CC client, we have a check to make sure that a file was within one
of the configured VOBs for the repository before including it in the
diff. This was hardcoding a `/` character, which didn't work on Windows
clients. This changes it to use `os.sep`.

While I was here I also updated the client name to include
"VersionVault" for better output in `rbt list-repo-types`.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12275/