Fix applying patches in Mercurial.
A change in RBTools 3.1 resulted in a duplicate `with_errors=` argument
passed to `execute()` in `apply_patch()`. `MercurialClient._execute()`
was explicitly setting this and also allowing any arbitrary `**kwargs`
to be passed in. This was missed due to lack of unit testing.

This changes `_execute()` to take in `with_errors` as an argument with
the default we want to set. It also adds unit testing for
`apply_patch()` to ensure this code path does not regress.

Testing Done:
Unit tests pass.

Tested by a customer that hit this issue.

Reviewed at https://reviews.reviewboard.org/r/12307/