Add a top-level doc for authenticating with RBTools.
RBTools authentication has never been explicitly documented (outside of
the small doc for `rbt login`). With the coming inclusion of SSO support
in Review Board, we need a much more visible set of instructions for
logging in with RBTools, especially with API tokens. This change adds
that to the manual.

Testing done:
Built HTML and checked the output.

Reviewed at https://reviews.reviewboard.org/r/12343/