Add a workflow document for VersionVault.
This adds a new page to the documentation going over the Review Board
workflow when using ClearCase or VersionVault. It duplicates a lot of
what we have in the `rbt post` document, but goes into additional
detail, and is written more for people who don't yet fully know how to
use Review Board. At the bottom is a full example of using Review Board
to review a UCM activity.

Testing Done:
- Spell checked new doc.
- Built HTML and checked output.

Reviewed at https://reviews.reviewboard.org/r/12344/