Fix rbt patch to not show byte strings during conflicts.
When `rbt patch` attempts to patch a change and encounters conflicts, it
displays the names of the affected files. This was being shown as a byte
string, rather than a Unicode string. We were doing the right thing by
converting it to a Unicode string for the JSON output, but not for the
displayed output.

This change converts the filename before outputting it for text or JSON.

Testing Done:
Triggered a patch with conflicts and saw the proper output.

Reviewed at https://reviews.reviewboard.org/r/12365/