Fix rbt post with --diff-filename, and improve file not found errors.
The support for `--diff-filename` regressed in RBTools 3.1, due to the
addition of a new `review_request_extra_data` attribute on
`SquashedDiff` that wasn't being provided (being a `namedtuple`, all
arguments always have to be provided). This is now set to the default of
`None`.

Along with this, the error message shown for a File Not Found contained
the filename but not the revision, making it hard to determine what was
missing. This is now included.

Testing Done:
Encountered both of these issues while testing another problem. Verified
that `--diff-filename` now works, and that the File Not Found error
contained the correct revision.

Reviewed at https://reviews.reviewboard.org/r/12366/