Fix running Bazaar unit tests with only Breezy installed.
RBTools was updated to support Breezy, but the unit tests were not. If
Breezy was installed as `brz`, but without a symlink to `bzr`, the unit
tests would be skipped.

This change fixes up the test suite to determine the correct version
(using `check_install` to avoid issues with stale `pyenv` shims or
"helpful" Linux instruction errors) and to run the correct one.

It also fixes up all calls to ensure a user/e-mail is set when operating
on a repository, avoiding errors that can be raised by Breezy.

Last, it avoids skipping when setting up the class-wide repository
state. This is to ensure that we don't skip during class setup, and soon
to ensure that each individual test has control over if it will skip.

Testing Done:
Ran unit tests with `bzr` and `brz` installed, with just `brz`
installed, with just `bzr` installed, and with neither installed.

Reviewed at https://reviews.reviewboard.org/r/12493/