Update the Mercurial unit tests to use build_client() and clean up tests.
This updates the Mercurial unit tests to build a `MercurialClient` in
each test where it's needed, rather than in `setUp()`. This is in
prepration for new tests that will need to handle client construction
differently.

This also cleans up a lot of the assertions to be less verbose and to be
more comprehensive.

We had a lot of combinations of checking instance type, key presence,
and then values, but this is far less useful (less accurate) than just
checking dictionaries for equality. We missed things because we weren't
doing these checks.

Given the work being done already in each function, I felt this was the
best time to clean this code up. It's caught some missing state we
weren't checking.

Testing Done:
Mercurial unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12508/