Update the Subversion unit tests to use build_client() and clean up tests.
This updates the Subversion unit tests to build a `SCNClient` in each
test where it's needed, rather than in `setUp()`. This is in prepration
for new tests that will need to handle client construction differently.

This also cleans up a lot of the assertions to be less verbose and to be
more comprehensive.

We had a lot of combinations of checking instance type, key presence,
and then values, but this is far less useful (less accurate) than just
checking dictionaries for equality. We missed things because we weren't
doing these checks.

It also fixes up diff checking to compare for content and not just a
hash, for each version of Subversion where we previously compared
hashes. This will be important for work on supporting alternative diff
tools.

Testing Done:
Subversion unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12511/