Improve output when scanning for a usable client.
This rewords some of the output when trying to scan for a usable client
but failing to find anything. It aims to add context to the messages,
helping to see the input (repository type or URL) that led to the error,
and to add instructions to help resolve the issue (saying to check for a
.reviewboardrc, what it should contain, and pointing users to
`rbt setup-repo`).

The work here is just a base for upcoming work on this logic to clean
things up and to provide error reporting that works with the upcoming
formal `SCMClient` dependency support.

Testing Done:
Tested all the error/warning conditions, verifying the output.

Reviewed at https://reviews.reviewboard.org/r/12512/