Move base rbtools.clients support into rbtools.clients.base.
We have a lot of support code in `rbtools.clients.__init__`, with more
coming. This has been hard to maintain, and was going to create some
circular dependency issues.

This cleanup has been long-overdue. This change moves `SCMClient`,
`RepositoryInfo`, `PatchAuthor`, and `PatchResult` into sub-modules of
`rbtools.clients.base`.

In the process, `SCMClient` has been renamed to `BaseSCMClient`, to
better fit modern naming of base classes.

There is a legacy `SCMClient` class in `rbtools.clients` that emits a
deprecation warning when instantiated. It will be removed in 5.0.

Aside from that, all the changes are mechanical or updates to references
in documentation.

Testing Done:
All unit tests pass.

Posted this change using this build.

Reviewed at https://reviews.reviewboard.org/r/12518/