Add a code reference for RBTools.
Despite the fact that RBTools provides APIs for talking to Review Board
and implementing custom commands and clients, we haven't documented the
codebase publicly.

This adds a code reference to the documentation's API section, and fixes
some syntax errors in the docs that were found during generation. We're
now using Sphinx 5.0, and beanbag-docutils 2.0, to build the
documentation.

There's a lot that needs to be done to clean up documentation in the
codebase, but that's work that can be done over time.

Testing Done:
Built the docs. There are still warnings, but they'd require more
extensive work.

Reviewed at https://reviews.reviewboard.org/r/12519/