Fix small issues with TFS diffs and add unit tests.
This change adds diff unit tests for all the TFS wrappers, helping
ensure we have as close to comprehensive diff coverage as possible. This
will be important for upcoming work on diff tools, and just ensures we
don't have any type-related issues.

This did uncover that we were fetching content for binary files when we
didn't need to. These were being fetched and then ignored. These code
paths have been updated to avoid the fetch for binary files.

The `rb-tfs` wrapper performs its own diffing, so we can't test or even
work with any of that. Ideally, that would be updated to provide
information we could use to generate a diff internally, but that's far
outside the scope of the work for RBTools 4.

Testing Done:
Unit tests pass on Python 3.7-3.11.

Reviewed at https://reviews.reviewboard.org/r/12568/