Don't crash if ClearCase/VersionVault directory changes include symlinks.
ClearCase and VersionVault can track symbolic links, which don't seem to show up
as elements of their own, but do show up in directory diffs. The way these were
represented (as `filename1 -> filename2`) was then crashing when trying to
query the OID.

This change wraps that section of our directory processing in a try/except, so
we can log the error and continue on. The result of this is that changes which
include added or deleted symlinks will post without errors. The resulting diff
will show the added line in the directory listing.

Testing Done:
Posted activity changes which involved symbolic links. Saw the expected diffs
instead of having `rbt post` crash.

Reviewed at https://reviews.reviewboard.org/r/12624/