Fix file renames/deletes and filenames with whitespace in git-p4 diffs.
Previously, the filename was gathered from the `diff --git ... line`,
but this is problematic in the case where the filename has spaces. The
parser was simply splitting on space.

The solution is to get the filename from the `--- ...` line, resetting
when we see the `diff ...` line. This is simpler as we can assume
anything following the `---` is the filename.  We also ensure that we
correctly identify file renames by reading the old filename and new
filename from these lines.

Further we include the headers for renames as generated by the Perforce
RBTools client (Moved from: and Moved to:). There's one special case
here where previously a 100% rename (with no diff) would silently fail
to send anything to the server. In this case, we now send something the
server recognises as a file rename operation, special casing like we do
in the perforce.py client.

Testing Done:
In a git-p4 repo with files containing spaces:
1. edit one file
2. rename one file
3. create a new file with space in it
4. delete a file with space in it

I compared the generated diffs from a pure Perforce repo and a git-p4
repo of the same location and ensured that we generate the same for
git-p4 as we do for Perforce on `rbt post`.

Reviewed at https://reviews.reviewboard.org/r/12258/